add days

+++ pojo/src/main/java/com/stanzaliving/booking/dto/response/AdvanceRentalDto.java
@@ -19,7 +19,7 @@ public class AdvanceRentalDto implements Serializable {
 
     private LocalDate endDate;
 
-    private Double advanceRentalMonths;
+    private int advanceRentalMonths;
 
     private Double advanceRentalMonthsAmount;
